fix drawn item index in event, fix input payment mint

+++ clients/js/src/anchorIdls/mallow_gumball.ts
@@ -1495,7 +1495,7 @@ export type MallowGumball = {
         },
         {
           name: 'index';
-          type: 'u64';
+          type: 'u32';
           index: false;
         }
       ];
@@ -3306,7 +3306,7 @@ export const IDL: MallowGumball = {
         },
         {
           name: 'index',
-          type: 'u64',
+          type: 'u32',
           index: false,
         },
       ],
